Cria cadastro de veiculo na pagina de oferecer carona
